requirements.txt
Pillow==9.0.0
opencv-python==4.2.0.34
fastapi==0.93.0
numpy==1.21.1
pycocotools==2.0.2
requests==2.26.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
typing_extensions==4.2.0
uvicorn==0.20.0